Changes argument parsing.
For mn and ex, there is no difference between
rigid 4.1 and 4.1B like there is for re. So, to
avoid having duplicate files on S3 with different
names, we convert 4.1B -> 4.1.

For ex, there is no difference between 4.0 and 4.1
so for the same reasons, we convert 4.1 -> 4.0.

We decided that having a non-default preset for
i2t would be on a case-by-case basis. As such,
there is no longer a need for it in the filename.