Bug fixes for using third-party tools.
In tools.py, when appending to the PATH via os.environ, had to add a
os.pathsep. Also fixed a typo in the mingw path so it's i686- not i6860.
In run_command, set the env keyword to os.environ to get the child
process to know about the path additions. This also required setting
shell=True.